Add a development container file
Development containers allow to develop the project in virtualised
environment, like on desktop with Docker, or remotely with providers
like GitHub Codespaces.

This makes it extremely easy for outsiders to contribute to the project,
since the environment is already set up for them, from Rust to the full
desktop environment to test Theseus GUI. It also makes project more
accessible to contributors who don't have the best hardware to do Rust
development.