change app category to 'Game'
As of now, the Modrinth launcher is listed in the category 'Internet' (alongside Firefox and Chrome), and not in 'Games'.
I’m on KDE Plasma, but since it’s labeled as 'Internet' in `modrinth-app.desktop` I assume this happens on every Linux.
This PR *changes* 'Entertainment' to 'Game'. I don’t know if the 'Entertainment' is needed for MacOS or Windows.