Create @modrinth/i18n package
@modrinth/i18n will be a package that contains the common messages that
are used by two or more packages, as well as utilities for working with
localisation, such as `defineMessage` macro and locale files resolver
abstraction.

It is pre-built in a manner that allows for harsh tree-shaking, meaning
the messages that will not be used, should not be included by the
bundlers.

All messages are also defined using the `defineMessage` macro that
cleans the descriptors and pre-parses `defaultMessage` to AST, which
serves both as validation, as well as allows its use as fallback even in
environments devoid of bundled-in compiler. When all other code parts
are switched to use this macro, this might allow to exclude messages
files for English language, potentially saving on removed duplication.

Please note that currently `defineMessages` does not include
`defaultMessage` because it is expected that consumers of
`@modrinth/i18n` will include and pre-compile extracted messages file.
This can be overriden by setting `MODRINTH_I18N_DEFAULT_MESSAGE`
environment variable to `1`.